feat(admin): fold Bulk Photo Upload + Map Heroes into the Photos tab
Removes two more redundant sidebar entries.

- Photos page header gets a "Bulk Upload" button (opens the existing
  bulk-photo route).
- Map hero backfill (coverage stats + runner) is now a MapHeroesSection
  rendered on the Photos tab; the standalone /admin/map-heroes route
  redirects there.
- Sidebar drops the "Bulk Photo Upload" and "Map Heroes" items.

Photos page test stubs the async MapHeroesSection (its logic is covered by
MapHeroBackfillRunner's own tests). Full suite green (2266).
