+++ src/components/admin/AdminNav.tsx
@@ -10,13 +10,11 @@ import {
   Camera,
   ClipboardList,
   Home,
-  Image as ImageIcon,
   KeyRound,
   LayoutDashboard,
   MapPin,
   Menu,
   MessageSquare,
-  Upload,
   Users,
   X,
   type LucideIcon,
@@ -37,12 +35,10 @@ const NAV_ITEMS: NavItem[] = [
   { href: "/admin/ai-research", label: "AI Research", icon: BrainCircuit },
   { href: "/admin/parks", label: "Parks", icon: MapPin },
   { href: "/admin/photos", label: "Photos", icon: Camera },
-  { href: "/admin/photos/bulk-upload", label: "Bulk Photo Upload", icon: Upload },
   { href: "/admin/conditions", label: "Trail Conditions", icon: Activity },
   { href: "/admin/reviews", label: "Reviews", icon: MessageSquare },
   { href: "/admin/claims", label: "Park Claims", icon: ClipboardList },
   { href: "/admin/operators", label: "Park Operators", icon: Building2 },
-  { href: "/admin/map-heroes", label: "Map Heroes", icon: ImageIcon },
   { href: "/admin/users", label: "Users", icon: Users },
   { href: "/admin/pre-grants", label: "Pre-grants", icon: KeyRound, superAdminOnly: true },
 ];
